refactor: input 받아오는 패턴 리팩토링
